requirements.txt
contourpy==1.1.0
easydict==1.10
fair-esm==2.0.0
fonttools==4.42.0
fsspec==2023.6.0
huggingface-hub==0.16.4
importlib-resources==6.0.0
matplotlib==3.7.2
ninja==1.11.1
packaging==23.1
pyparsing==3.0.9
python-graphviz==0.20.1
pyyaml==6.0.1
rdkit-pypi==2022.9.5
regex==2023.8.8
safetensors==0.3.2
scipy==1.11.1
tokenizers==0.13.3
torchviz==0.0.2
transformers==4.32.0
zipp==3.16.2